+++ app/controllers/docs_controller.rb
@@ -2,6 +2,26 @@
 
 # Displays publicly available documentation.
 class DocsController < ApplicationController
+  # GET /docs/adverse-events
+  def adverse_events
+  end
+
+  # GET /docs/roles
+  def roles
+  end
+
+  # GET /docs/notifications
+  def notifications
+  end
+
+  # GET /docs/blinding
+  def blinding
+  end
+
+  # GET /docs/sites
+  def sites
+  end
+
   # GET /docs/technical
   def technical
   end